requirements.txt
fastapi==0.115.2
uvicorn==0.32.0
jinja2==3.1.4
python-multipart==0.0.12
emails==0.6
python-dotenv==1.0.1